Fix Runaway portnames and listing of internal virtual ports on LInux (#17)
* gitignore - ignore Emacs tmp files

* Port Listings - Linux -  fix runaway portnames and listing of internal ports

On Linux RtMidi appears to create a new virtual MIDI port per physical input or output port you wish to open for IO.  This has two negative connotations on Linux:

1. in the GUI, these internal virtual MIDI ports are listed which may confuse users
2. when the user hotswaps a MIDI device, the update trigger scans all new MIDI devices which now includes the newly created MIDI ports (they weren't listed previously as the first scan happened before any MIDI ports were opened). As the virtual MIDI ports were listed, they were automatically connected to (as is also the case with any new physical MIDI device) which then created more additional virtual devices which then got listed in the next scan and opened triggering more virtual devices etc.

This patch targets both of these issues.

A new pair of MIDI input name listing functions have been created. These act similarly to `getInputNames()` and `getOutputNames()` but they also filter out any names that start with `rtmidi_` which are assumed to be one of the virtual ports described above.  These new functions are used for scanning for ports to open which therefore means they won't be automatically connected to - therefore not creating a cascading set of connections and new virtual ports.

We also use these functions to provide the port names up to the Erlang layer. This stops the GUI from listing the internally created virtual ports.

Note: I'm absolutely not a C++ expert - so it may very well be the case that I'm breaking some rule about memory safety and I've introduced a memory leak with an every growing heap of vectors of strings - so this patch should definitely be carefully scrutinized.

This patch also assumes that no realistic port name on any platform starts with the prefix `rtmidi_`. If this stops being the case then this needs to be addressed.